refactor(shmem): separate host/device types to eliminate ibverbs from device compilation path
