+++ include/mori/application/memory/symmetric_memory.hpp
@@ -21,6 +21,9 @@
 // SOFTWARE.
 #pragma once
 
+// VMMChunkKey, HeapType, SymmMemObj, SymmMemObjPtr, RdmaMemoryRegion are defined in
+// application_device_types.hpp so that device (HIP/CUDA) compilation units can include
+// them without pulling in STL or ibverbs headers.
 #include <hip/hip_runtime_api.h>
 #include <linux/types.h>
 #include <stdint.h>
@@ -29,6 +32,7 @@
 #include <unordered_map>
 #include <vector>
 
+#include "mori/application/application_device_types.hpp"
 #include "mori/application/bootstrap/bootstrap.hpp"
 #include "mori/application/context/context.hpp"
 #include "mori/application/memory/va_manager.hpp"
@@ -38,91 +42,6 @@
 
 namespace mori {
 namespace application {
-
-// Heap type for memory allocation
-enum class HeapType {
-  Normal,   // Normal cached memory
-  Uncached  // Uncached memory
-};
-
-struct VMMChunkKey {
-  uint32_t key;         // RDMA lkey or rkey
-  uintptr_t next_addr;  // Address of next chunk boundary (for calculating chunk_size)
-
-  VMMChunkKey() : key(0), next_addr(0) {}
-  VMMChunkKey(uint32_t k, uintptr_t addr) : key(k), next_addr(addr) {}
-};
-
-struct SymmMemObj {
-  void* localPtr{nullptr};
-  uintptr_t* peerPtrs{nullptr};
-  uintptr_t* p2pPeerPtrs{nullptr};
-  size_t size{0};
-  // For Rdma
-  uint32_t lkey{0};
-  uint32_t* peerRkeys{nullptr};
-
-  // For VMM allocations: chunk key information (nvshmem-style)
-  // vmmLkeyInfo[i] contains lkey and next_addr for chunk i
-  // vmmRkeyInfo[i * worldSize + pe] contains rkey and next_addr for chunk i, PE pe
-  VMMChunkKey* vmmLkeyInfo{nullptr};
-  VMMChunkKey* vmmRkeyInfo{nullptr};
-  size_t vmmNumChunks{0};  // Total number of chunks in VMM heap
-  int worldSize{0};
-  // For IPC
-  hipIpcMemHandle_t* ipcMemHandles{nullptr};  // should only placed on cpu
-
-  // For Sdma
-  anvil::SdmaQueueDeviceHandle** deviceHandles_d = nullptr;  // should only placed on GPU
-  HSAuint64* signalPtrs = nullptr;                           // should only placed on GPU
-  uint32_t sdmaNumQueue = 2;                                 // number of sdma queue
-  HSAuint64* expectSignalsPtr = nullptr;                     // should only placed on GPU
-  // Remote signal: peerSignalPtrs[pe] points to PE pe's signalPtrs mapped into local address space.
-  // SdmaPutThread writes ATOMIC to peerSignalPtrs[remotePe] + myPe*sdmaNumQueue + qId,
-  // so the remote PE can directly read its own signalPtrs to detect completion.
-  HSAuint64** peerSignalPtrs = nullptr;  // should only placed on GPU
-
-  __device__ __host__ RdmaMemoryRegion GetRdmaMemoryRegion(int pe) const {
-    RdmaMemoryRegion mr;
-    mr.addr = peerPtrs[pe];
-    mr.lkey = lkey;
-    mr.rkey = peerRkeys[pe];
-    mr.length = size;
-    return mr;
-  }
-
-  // Get pointers
-  inline __device__ __host__ void* Get() const { return localPtr; }
-  inline __device__ __host__ void* Get(int pe) const {
-    return reinterpret_cast<void*>(p2pPeerPtrs[pe]);
-  }
-
-  template <typename T>
-  inline __device__ __host__ T GetAs() const {
-    return reinterpret_cast<T>(localPtr);
-  }
-  template <typename T>
-  inline __device__ __host__ T GetAs(int pe) const {
-    return reinterpret_cast<T>(p2pPeerPtrs[pe]);
-  }
-};
-
-struct SymmMemObjPtr {
-  SymmMemObj* cpu{nullptr};
-  SymmMemObj* gpu{nullptr};
-
-  bool IsValid() { return (cpu != nullptr) && (gpu != nullptr); }
-
-#if defined(__HIPCC__) || defined(__CUDACC__)
-  __host__ SymmMemObj* operator->() { return cpu; }
-  __device__ SymmMemObj* operator->() { return gpu; }
-  __host__ const SymmMemObj* operator->() const { return cpu; }
-  __device__ const SymmMemObj* operator->() const { return gpu; }
-#else
-  SymmMemObj* operator->() { return cpu; }
-  const SymmMemObj* operator->() const { return cpu; }
-#endif
-};
 
 class SymmMemManager {
  public:
