+++ include/mori/core/transport/rdma/primitives.hpp
@@ -21,146 +21,21 @@
 // SOFTWARE.
 #pragma once
 
-#include <limits.h>
-#include <stdint.h>
-
 #include "infiniband/verbs.h"
+#include "mori/core/transport/rdma/core_device_types.hpp"
 
 namespace mori {
 namespace core {
 
-enum ProviderType {
-  Unknown = 0,
-  // Mellanox direct verbs
-  MLX5 = 1,
-  // Broadcom direct verbs
-  BNXT = 2,
-  // Pensando direct verbs
-  PSD = 3,
-  // Ib verbs
-  IBVERBS = 4,
-};
-
-typedef enum {
-  AMO_ACK = 1,
-  AMO_INC,
-  AMO_SET,
-  AMO_ADD,
-  AMO_AND,
-  AMO_OR,
-  AMO_XOR,
-  AMO_SIGNAL,
-  SIGNAL_SET,
-  SIGNAL_ADD,
-  AMO_SIGNAL_SET = SIGNAL_SET,
-  AMO_SIGNAL_ADD = SIGNAL_ADD,
-  AMO_END_OF_NONFETCH,
-  AMO_FETCH,
-  AMO_FETCH_INC,
-  AMO_FETCH_ADD,
-  AMO_FETCH_AND,
-  AMO_FETCH_OR,
-  AMO_FETCH_XOR,
-  AMO_SWAP,
-  AMO_COMPARE_SWAP,
-  AMO_OP_SENTINEL = INT_MAX,
-} atomicType;
-
-#define OUTSTANDING_TABLE_SIZE (65536)
-struct WorkQueueHandle {
-  uint32_t postIdx{0};     // numbers of wqe that post to work queue
-  uint32_t dbTouchIdx{0};  // numbers of wqe that touched doorbell
-  uint32_t doneIdx{0};     // numbers of wqe that have been consumed by nic
-  uint32_t readyIdx{0};
-  union {
-    struct {
-      uint32_t msntblSlotIdx;
-      uint32_t psnIdx;  // for bnxt msn psn index calculate
-    };
-    uint64_t msnPack{0};
-  };
-  void* sqAddr{nullptr};
-  void* msntblAddr{nullptr};  // for bnxt
-  void* rqAddr{nullptr};
-  void* dbrRecAddr{nullptr};
-  void* dbrAddr{nullptr};
-  void* rqdbrAddr{nullptr};
-  uint32_t mtuSize{4096};
-  uint32_t sqWqeNum{0};
-  uint32_t msntblNum{0};
-  uint32_t rqWqeNum{0};
-  uint32_t postSendLock{0};
-  uint64_t outstandingWqe[OUTSTANDING_TABLE_SIZE]{0};
-  bool color;
-  uint64_t sq_dbval{0};
-  uint64_t rq_dbval{0};
-};
-
-struct CompletionQueueHandle {
-  void* cqAddr{nullptr};
-  void* dbrRecAddr{nullptr};
-  void* dbrAddr{nullptr};
-  uint32_t consIdx{0};      // numbers of cqe that have been completed
-  uint32_t needConsIdx{0};  // numbers of cqe that should be consumed
-  uint32_t activeIdx{0};    // numbers of cqe that under processing but not completed
-  uint32_t cq_consumer{0};
-  uint32_t cq_dbpos{0};
-  uint32_t cqeNum{0};
-  uint32_t cqeSize{0};
-  uint32_t pollCqLock{0};
-  uint64_t cq_dbval{0};
-};
-
+// IBVerbsHandle is host-only: ibv_qp/ibv_cq/etc. are ibverbs objects managed on the host.
+// Device code must not access these fields. For device-side RDMA, use WorkQueueHandle /
+// CompletionQueueHandle instead (defined in core_device_types.hpp).
 struct IBVerbsHandle {
   ibv_qp* qp{nullptr};
   ibv_cq* cq{nullptr};
   ibv_srq* srq{nullptr};
   ibv_comp_channel* compCh{nullptr};
 };
-
-struct IbufHandle {
-  uintptr_t addr{0};
-  uint32_t lkey{0};
-  uint32_t rkey{0};
-  uint32_t nslots{0};
-  uint32_t head{0};
-  uint32_t tail{0};
-};
-
-/* ---------------------------------------------------------------------------------------------- */
-/*                                        Utility Functions                                       */
-/* ---------------------------------------------------------------------------------------------- */
-#define BSWAP64(x)                                                                 \
-  ((((x) & 0xff00000000000000ull) >> 56) | (((x) & 0x00ff000000000000ull) >> 40) | \
-   (((x) & 0x0000ff0000000000ull) >> 24) | (((x) & 0x000000ff00000000ull) >> 8) |  \
-   (((x) & 0x00000000ff000000ull) << 8) | (((x) & 0x0000000000ff0000ull) << 24) |  \
-   (((x) & 0x000000000000ff00ull) << 40) | (((x) & 0x00000000000000ffull) << 56))
-
-#define BSWAP32(x)                                                                      \
-  ((((x) & 0xff000000) >> 24) | (((x) & 0x00ff0000) >> 8) | (((x) & 0x0000ff00) << 8) | \
-   (((x) & 0x000000ff) << 24))
-
-#define BSWAP16(x) ((((x) & 0xff00) >> 8) | (((x) & 0x00ff) << 8))
-
-#define HTOBE64(x) BSWAP64(x)
-#define HTOBE32(x) BSWAP32(x)
-#define HTOBE16(x) BSWAP16(x)
-
-#if BYTE_ORDER == LITTLE_ENDIAN
-#define BE16TOH(x) BSWAP16(x)
-#define BE32TOH(x) BSWAP32(x)
-#define BE64TOH(x) BSWAP64(x)
-#define LE16TOH(x) (x)
-#define LE32TOH(x) (x)
-#define LE64TOH(x) (x)
-#elif BYTE_ORDER == BIG_ENDIAN
-#define BE16TOH(x) (x)
-#define BE32TOH(x) (x)
-#define BE64TOH(x) (x)
-#define LE16TOH(x) BSWAP16(x)
-#define LE32TOH(x) BSWAP32(x)
-#define LE64TOH(x) BSWAP64(x)
-#endif
 
 }  // namespace core
 }  // namespace mori
