+++ include/mori/shmem/shmem_device_api.hpp
@@ -23,7 +23,7 @@
 
 #include <assert.h>
 
-#include "mori/application/application.hpp"
+#include "mori/application/application_device_types.hpp"
 #include "mori/core/core.hpp"
 #include "mori/shmem/internal.hpp"
 #include "mori/shmem/shmem_device_kernels.hpp"
